requirements.txt
alembic==1.4.2
aniso8601==6.0.0
astroid==2.2.5
attrs==19.3.0
Authlib==0.14.3
awscli==1.18.64
botocore==1.16.14
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==3.0
dnspython==2.0.0
docutils==0.15.2
ecdsa==0.15
email-validator==1.1.1
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.3.1
requests==2.24.0
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.4
typed-ast==1.3.5
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.1
zipp==3.1.0
